new improvements for ML

+++ pkg/simulator/src/pkg/model/ml-smart-plc-intent.go
@@ -7,7 +7,7 @@ type MLSmartPlacementIntent struct {
 
 // SpaceAPP (for single app)  : 1) Required mvCPU 2) required Memory 3) Required Latency 4) Current MEC 5) Current RAN
 type SpaceAPP struct {
-	AppCharacteristic [5]int64 `json:"app_characteristic"`
+	AppCharacteristic [5]int `json:"app_characteristic"`
 }
 
 type SpaceMECs struct {
